refacto: Add InteractionButton and use it in ControlItem and radio and checkbox indicators
