+++ OUDS/Core/Components/Sources/Checkbox/Internal/CheckboxIndicatorModifiers.swift
@@ -16,36 +16,38 @@ import OUDSFoundations
 import OUDSTokensSemantic
 import SwiftUI
 
-// MARK: - Radio Indicator Style
+// MARK: - Checkbox Indicator Modifier
 
-/// A `ViewModier` to apply to the ``RadioIndicator`` component.
-/// It will define the look and feel of the indicator depending to the ``InteractionState`` and some flags
-struct RadioIndicatorStyle: ViewModifier {
+/// A `ViewModier` to apply to the ``CheckboxIndicator`` component.
+/// It will define the look and feel of the indicator depending to the ``InteractionState``,
+/// the ``OUDSCheckboxIndicatorState`` and if there is an error context or not.
+struct CheckboxIndicatorModifier: ViewModifier {
 
     // MARK: - Properties
 
     let interactionState: InteractionState
-    let isOn: Bool
+    let indicatorState: OUDSCheckboxIndicatorState
     let isError: Bool
 
     // MARK: - Body
 
     func body(content: Content) -> some View {
         content
-            .modifier(RadioIndicatorBorderModifier(interactionState: interactionState, isOn: isOn, isError: isError))
-            .modifier(RadioIndicatorForegroundModifier(interactionState: interactionState, isOn: isOn, isError: isError))
-            .modifier(RadioIndicatorBackgroundModifier(interactionState: interactionState, isOn: isOn, isError: isError))
+            .modifier(SizeFrameModifier())
+            .modifier(CheckboxIndicatorBorderModifier(interactionState: interactionState, indicatorState: indicatorState, isError: isError))
+            .modifier(CheckboxIndicatorForegroundModifier(interactionState: interactionState, indicatorState: indicatorState, isError: isError))
+            .modifier(CheckboxIndicatorBackgroundModifier(interactionState: interactionState, indicatorState: indicatorState, isError: isError))
     }
 }
 
-// MARK: - Radio Indicator Foreground Modifier
+// MARK: - Checkbox Indicator Foreground Modifier
 
-private struct RadioIndicatorForegroundModifier: ViewModifier {
+private struct CheckboxIndicatorForegroundModifier: ViewModifier {
 
     // MARK: - Properties
 
     let interactionState: InteractionState
-    let isOn: Bool
+    let indicatorState: OUDSCheckboxIndicatorState
     let isError: Bool
 
     @Environment(\.theme) private var theme
@@ -77,7 +79,12 @@ private struct RadioIndicatorForegroundModifier: ViewModifier {
         if isError {
             return theme.colors.colorActionNegativeEnabled
         } else {
-            return isOn ? theme.colors.colorActionSelected : theme.colors.colorActionEnabled
+            switch indicatorState {
+            case .selected, .indeterminate:
+                return theme.colors.colorActionSelected
+            case .unselected:
+                return theme.colors.colorActionEnabled
+            }
         }
     }
 
@@ -91,21 +98,21 @@ private struct RadioIndicatorForegroundModifier: ViewModifier {
 
     private var disabledColor: MultipleColorSemanticTokens {
         guard !isError else {
-            OL.fatal("An OUDSRadio with a disabled state and an error situation has been detected, which is not allowed."
+            OL.fatal("An OUDS Checkbox with a disabled state and an error situation has been detected, which is not allowed."
                      + " Only non-error situation are allowed to have a disabled state.")
         }
         return theme.colors.colorActionDisabled
     }
 }
 
-// MARK: - Radio Indicator Background Modifier
+// MARK: - Checkbox Indicator Background Modifier
 
-private struct RadioIndicatorBackgroundModifier: ViewModifier {
+private struct CheckboxIndicatorBackgroundModifier: ViewModifier {
 
     // MARK: - Properties
 
     let interactionState: InteractionState
-    let isOn: Bool
+    let indicatorState: OUDSCheckboxIndicatorState
     let isError: Bool
 
     @Environment(\.theme) private var theme
@@ -138,30 +145,30 @@ private struct RadioIndicatorBackgroundModifier: ViewModifier {
     }
 
     private var hoverColor: Color {
-        Color.clear
+        theme.controlItem.controlItemColorBgHover.color(for: colorScheme)
     }
 
     private var pressedColor: Color {
-        Color.clear
+        theme.controlItem.controlItemColorBgPressed.color(for: colorScheme)
     }
 
     private var disabledColor: Color {
         guard !isError else {
-            OL.fatal("An OUDSRadio with a disabled state and an error situation has been detected, which is not allowed."
+            OL.fatal("An OUDS Checkbox with a disabled state and an error situation has been detected, which is not allowed."
                      + " Only non-error situation are allowed to have a disabled state.")
         }
         return Color.clear
     }
 }
 
-// MARK: - Radio Indicator Border Modifier
+// MARK: - Checkbox IndicIndicatorator Border Modifier
 
-private struct RadioIndicatorBorderModifier: ViewModifier {
+private struct CheckboxIndicatorBorderModifier: ViewModifier {
 
     // MARK: - Properties
 
     let interactionState: InteractionState
-    let isOn: Bool
+    let indicatorState: OUDSCheckboxIndicatorState
     let isError: Bool
 
     @Environment(\.theme) private var theme
@@ -196,7 +203,12 @@ private struct RadioIndicatorBorderModifier: ViewModifier {
         if isError {
             return theme.colors.colorActionNegativeEnabled
         } else {
-            return isOn ? theme.colors.colorActionSelected : theme.colors.colorActionEnabled
+            switch indicatorState {
+            case .selected, .indeterminate:
+                return theme.colors.colorActionSelected
+            case .unselected:
+                return theme.colors.colorActionEnabled
+            }
         }
     }
 
@@ -218,7 +230,7 @@ private struct RadioIndicatorBorderModifier: ViewModifier {
 
     private var disabledColor: MultipleColorSemanticTokens {
         guard !isError else {
-            OL.fatal("An OUDSRadio with a disabled state and an error situation has been detected, which is not allowed"
+            OL.fatal("An OUDS Checkbox with a disabled state and an error situation has been detected, which is not allowed"
                      + " Only non-error situation are allowed to have a disabled state.")
         }
         return theme.colors.colorActionDisabled
@@ -240,24 +252,57 @@ private struct RadioIndicatorBorderModifier: ViewModifier {
     }
 
     private var enabledWidth: CGFloat {
-        isOn ? theme.radioButton.radioButtonBorderWidthSelected : theme.radioButton.radioButtonBorderWidthUnselected
+        switch indicatorState {
+        case .selected, .indeterminate:
+            return theme.checkbox.checkboxBorderWidthSelected
+        case .unselected:
+            return theme.checkbox.checkboxBorderWidthUnselected
+        }
     }
 
     private var hoverWidth: CGFloat {
-        isOn ? theme.radioButton.radioButtonBorderWidthSelectedHover : theme.radioButton.radioButtonBorderWidthUnselectedHover
+        switch indicatorState {
+        case .selected, .indeterminate:
+            return theme.checkbox.checkboxBorderWidthSelectedHover
+        case .unselected:
+            return theme.checkbox.checkboxBorderWidthUnselectedHover
+        }
     }
 
     private var pressedWidth: CGFloat {
-        isOn ? theme.radioButton.radioButtonBorderWidthSelectedPressed : theme.radioButton.radioButtonBorderWidthUnselectedPressed
+        switch indicatorState {
+        case .selected, .indeterminate:
+            return theme.checkbox.checkboxBorderWidthSelectedPressed
+        case .unselected:
+            return theme.checkbox.checkboxBorderWidthUnselectedPressed
+        }
     }
 
     private var disabledWidth: CGFloat {
-        isOn ? theme.radioButton.radioButtonBorderWidthSelected : theme.radioButton.radioButtonBorderWidthUnselected
+        switch indicatorState {
+        case .selected, .indeterminate:
+            return theme.checkbox.checkboxBorderWidthSelected
+        case .unselected:
+            return theme.checkbox.checkboxBorderWidthUnselected
+        }
     }
 
     // MARK: - Border radius
 
     private var appliedBorderRadius: CGFloat {
-        theme.radioButton.radioButtonBorderRadius
+        theme.checkbox.checkboxBorderRadius
+    }
+}
+
+// MARK: - Size Frame Modifier
+
+private struct SizeFrameModifier: ViewModifier {
+
+    @Environment(\.theme) private var theme
+
+    func body(content: Content) -> some View {
+        content
+            .frame(width: theme.checkbox.checkboxSizeIndicator,
+                   height: theme.checkbox.checkboxSizeIndicator)
     }
 }
