+++ OUDS/Core/Components/Sources/Checkbox/OUDSCheckboxItemIndeterminate.swift
@@ -160,16 +160,11 @@ public struct OUDSCheckboxItemIndeterminate: View {
     // MARK: Body
 
     public var body: some View {
-        Button("") {
-            if !layoutData.isReadOnly {
-                $selection.wrappedValue.toggle()
-            }
-        }
-        .accessibilityRemoveTraits([.isButton]) // .isToggle trait for iOS 17+
-        .accessibilityLabel(a11yLabel(layoutData: layoutData))
-        .accessibilityValue(selection.a11yDescription.localized())
-        .accessibilityHint(a11yHint(isReadOnly: layoutData.isReadOnly, indicatorState: selection))
-        .buttonStyle(ControlItemStyle(indicatorType: .checkBox($selection), layoutData: layoutData))
+        ControlItem(indicatorType: .checkBox($selection), layoutData: layoutData)
+            .accessibilityRemoveTraits([.isButton]) // .isToggle trait for iOS 17+
+            .accessibilityLabel(a11yLabel(layoutData: layoutData))
+            .accessibilityValue(selection.a11yDescription.localized())
+            .accessibilityHint(a11yHint(isReadOnly: layoutData.isReadOnly, indicatorState: selection))
     }
 
     /// Forges a string to vocalize with *Voice Over* describing the component state.
