+++ OUDS/Core/Components/Sources/Internal/InteractionState/InteractionState.swift
@@ -29,12 +29,13 @@ enum InteractionState {
     /// The component is not disabled but user cannot interact with it still. Almost enabled.
     case readOnly // (╯°□°）╯︵ ┻━┻
 
-/// Create the state according to current states.
-/// - Parameters:
-///    - isEnabled: true if enabled, false otherwise (i.e. get for `isEnabled` environment)
-///    - isHover: true if a view is hoverd, false otherwise (i.e. get from `onHover`)
-///    - isPressed: true if pressed, false otherwise (i.e. get from `ButtonStyle.Configuration`)
-///    - isReadOnly: true if read only, false otherwise
+    /// Create the state according to flags.
+    ///
+    /// - Parameters:
+    ///    - isEnabled: true if enabled, false otherwise (i.e. get for `isEnabled` environment)
+    ///    - isHover: true if a view is hoverd, false otherwise (i.e. get from `onHover`)
+    ///    - isPressed: true if pressed, false otherwise (i.e. get from `ButtonStyle.Configuration`)
+    ///    - isReadOnly: true if read only, false otherwise
     init(isEnabled: Bool, isHover: Bool, isPressed: Bool, isReadOnly: Bool = false) {
         if isReadOnly {
             self = .readOnly
